Provide a maxSlippage parameter for the trade script (#13)
By trading randomly, particularly on xDAI we are doing a lot of bad trades with high slippage (💩💰). My first attempt was to groom the default token list to only allow for more "known" tokens but even there we cannot be guaranteed that there is good liquidity available (e.g. GNO has terrible liquidity on xDAI).

This PR therefore suggests to add an maxSlippage parameter to the execution to limit the trades we are doing to a maximum loss (1% by default). Note, that at the moment we can only estimate slippage as there is not yet an endpoint to get the current spot price of a pair.

While this reduces the token pairs on which we can realistically trade a lot, it will allow to have the bot run more autonomously (right now I have to refill it about once a day)

This should allow us to make bad trades for about a month (@ 1 trade/hour) before we need to refund the trading account.