Filter tokens with insufficient balance for fee (#14)
We are already checking for sufficient sell token balance to pay for the trading fee before placing the order. However, if we get to this point (e.g. because or random token pair picker chose a token for which we only have a breadcrumb amount) we fail the run and cause an alert.

This PR makes it so that these insufficient trade candidates are already filtered before making the random choice.

### Test Plan

Still able to run `yarn hardhat trade --network rinkeby`